Kandula-App/requirements.txt
dependency-injector
flask
bootstrap-flask
pygithub
pyyaml
pytest-flask
pytest-cov
docker
BeautifulSoup4
moto
json-logging
flake8
flake8-colors
yamllint
psycopg2-binary
python-dotenv

# boto3==1.26.160
# botocore==1.29.160
# jmespath==1.0.1
# psycopg2==2.9.6
# psycopg2-binary==2.9.6
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# s3transfer==0.6.1
# six==1.16.0
# urllib3==1.26.16